Accessor and mutator method for imageGalleryId and imageProfileId https://app.asana.com/0/1108331976344856/1108331976344860
